% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{add_legendgrouptitle}
\alias{add_legendgrouptitle}
\title{Add legend group titles}
\usage{
add_legendgrouptitle(p, traceInfo = NULL, ...)
}
\arguments{
\item{p}{a plotly plot.}

\item{traceInfo}{a data frame with legendgroup (i.e. title) and trace (i.e. trace number) columns. Default=NULL will be supplied by calling get_traceInfo(p). `traceInfo$legendgroup` is a factor whose level sequence determines legend group sequence.}

\item{...}{other parameters passed to all `style(legendgrouptitle=list(...),)`.}
}
\value{
a plotly plot
}
\description{
Add legend group titles
}
