% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdriver.R
\name{apptest}
\alias{apptest}
\title{Start to serve web app at temp/ and lauch chromedriver for inspection}
\usage{
apptest()
}
\value{
a session
}
\description{
Start to serve web app at temp/ and lauch chromedriver for inspection
}
