% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{attach_io_fromClipboardUI}
\alias{attach_io_fromClipboardUI}
\title{Get all input/output ids from UI code in the clipboard and attach input output lists for typing double check}
\usage{
attach_io_fromClipboardUI()
}
\value{
attach an environment of input/output
}
\description{
Get all input/output ids from UI code in the clipboard and attach input output lists for typing double check
}
