% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_restyle.R
\name{checkTraceColumns}
\alias{checkTraceColumns}
\title{check trace columns from a given plotly plot}
\usage{
checkTraceColumns(plt)
}
\arguments{
\item{plt}{a plotly plot}
}
\value{
a data frame of trace feature columns
}
\description{
check trace columns from a given plotly plot
}
