% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{do_add_trace}
\alias{do_add_trace}
\title{do.call on plotly add_trace}
\usage{
do_add_trace(p, .trace)
}
\arguments{
\item{p}{a plotly object}

\item{.trace}{a list of argument=value pairs that defines a trace. Can be a plotly object$x$data[[i]] value.}
}
\description{
do.call on plotly add_trace
}
