% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{do_config}
\alias{do_config}
\title{do.call on plotly config}
\usage{
do_config(p, .config)
}
\arguments{
\item{p}{a plotly object}

\item{.config}{a list of argument=value pairs that defines config. Can be a plotly object$x$config value.}
}
\description{
do.call on plotly config
}
