% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{do_layout}
\alias{do_layout}
\title{do.call on plotly layout}
\usage{
do_layout(p, .layout)
}
\arguments{
\item{p}{a plotly object}

\item{.layout}{a list of argument=value pairs that defines layout. Can be a plotly object$x$layout value.}
}
\description{
do.call on plotly layout
}
