% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{fromClipJSON}
\alias{fromClipJSON}
\title{Import clipboard of Plotly Studio json view as a list}
\usage{
fromClipJSON()
}
\description{
Import clipboard of Plotly Studio json view as a list
}
\examples{
none.
}
