% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_manual.R
\name{generate_find_attribute}
\alias{generate_find_attribute}
\title{Generate find_attribute function under given plotly reference url}
\usage{
generate_find_attribute(url, isLayoutSection = T)
}
\arguments{
\item{url}{a plotly reference url, such as https://plotly.com/r/reference/scatter/}
}
\value{
a webpage with all the result listing found.
}
\description{
Generate find_attribute function under given plotly reference url
}
\examples{
find_attr <- generate_find_attribute("https://plotly.com/r/reference/scatter/")
}
