% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_inputNamesDefaults}
\alias{get_inputNamesDefaults}
\title{Get input arguments' names and their defaults}
\usage{
get_inputNamesDefaults(fn)
}
\arguments{
\item{fn}{a function}
}
\value{

}
\description{
Get input arguments' names and their defaults
}
\examples{
none.
}
