% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{get_legendGroupTitles}
\alias{get_legendGroupTitles}
\title{Get legend group titles in trace orders}
\usage{
get_legendGroupTitles(df, label_by, group_by)
}
\arguments{
\item{df}{a dataframe}

\item{label_by}{a name. The column of df that is used to generate legend label}

\item{group_by}{a name. The column of df that is used for legendgroup.}
}
\value{
a character vector of legendgrouptitles in trace order.
}
\description{
Get legend group titles in trace orders
}
