% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{get_ranked_splitNames}
\alias{get_ranked_splitNames}
\title{Get ranked split names}
\usage{
get_ranked_splitNames(.plot)
}
\arguments{
\item{.plot}{a plotly plot}
}
\value{

}
\description{
Get ranked split names
}
\examples{
none.
}
