% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_restyle.R
\name{get_restyleInfo}
\alias{get_restyleInfo}
\title{Title}
\usage{
get_restyleInfo(plt0, plt1, traceId)
}
\arguments{
\item{plt0}{current plotly plot}

\item{plt1}{new plotly plot}

\item{traceId}{the name of one plotly_build data column that identifies trace. If not sure, checkTraceColumns(plt0) will give you all the available columns across traces.}
}
\value{
a list of two, one tell which traces have changed, the other give the restyle content of each changed trace.
}
\description{
Title
}
