% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{get_traceInfo}
\alias{get_traceInfo}
\title{Get trace information: its name, trace number and legend group}
\usage{
get_traceInfo(p)
}
\arguments{
\item{p}{a plotly plot.}
}
\value{
a data frame
}
\description{
Get trace information: its name, trace number and legend group
}
