% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{getAttributeListPairs}
\alias{getAttributeListPairs}
\title{Get attribute list pairs from a trace list}
\usage{
getAttributeListPairs(traceInf, attr2keep)
}
\arguments{
\item{traceInf}{a expression of trace list}

\item{attr2keep}{a character of attributes to keep}
}
\value{
an invisible character expression of the list pair. Implicitly copy to clipboard.
}
\description{
Get attribute list pairs from a trace list
}
