% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{guide_color2}
\alias{guide_color2}
\title{Generate diverging two party discrete color guide}
\usage{
guide_color2(parties, colorcodes, numbers)
}
\arguments{
\item{parties}{a list of two characters}

\item{colorcodes}{a list of two character vectors with equal length. Each element value represents the discrete colors of a party}

\item{numbers}{the cut numbers. If each party has 5 colors, then there will be 6 cut numbers.}
}
\value{
a html tag
}
\description{
Generate diverging two party discrete color guide
}
