% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{hex2rgba}
\alias{hex2rgba}
\title{Convert 6 digit hex color into a statement of "rgba(r, g, b, a)"}
\usage{
hex2rgba(hexcolor, alpha = 1)
}
\arguments{
\item{hexcolor}{a character vector of 6 digits hex}

\item{alpha}{numeric, between 0 and 1. Default=1}
}
\value{

}
\description{
Convert 6 digit hex color into a statement of "rgba(r, g, b, a)"
}
\examples{
hex2rgba("#294968", "#217BBB")
}
