% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{Layout}
\alias{Layout}
\title{Create a layout instance}
\usage{
Layout()
}
\value{
a layout instance ly. ly$layout(p, ..)  to be used as plotly::layout. ly$layout(...) simply creates a list of layout specifications. ly$layout(..., lytitle=NULL) where lytitle is a decription of this layout when ly$.layout shows all the layouts lytitle sill be element names. ly$mergedLayout() merge all layouts into one.
}
\description{
Create a layout instance
}
