% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{merge_list}
\alias{merge_list}
\title{Merge two list while respecting original list structure}
\usage{
merge_list(...)
}
\arguments{
\item{...}{list arguments}
}
\value{
a merged list
}
\description{
Merge two list while respecting original list structure
}
