% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_dependency.R
\name{moveDep_out}
\alias{moveDep_out}
\title{Move dependency function out of R scripts to dependency.R in /R}
\usage{
moveDep_out(filepath)
}
\arguments{
\item{filepath}{R script filepath}
}
\description{
Move dependency function out of R scripts to dependency.R in /R
}
\examples{
none
}
