% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{prepare_githubPage}
\alias{prepare_githubPage}
\title{prepare for github page publication}
\usage{
prepare_githubPage(tag)
}
\arguments{
\item{tag}{a tag object}
}
\value{

}
\description{
prepare for github page publication
}
\examples{
none
}
