% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quo_as.is}
\alias{quo_as.is}
\title{From a character of colname, generate its I(objectName) expression in quo}
\usage{
quo_as.is(colname)
}
\arguments{
\item{colname}{a character of variable name, such as "y2012".}
}
\value{
its quo(I(y2012))
}
\description{
From a character of colname, generate its I(objectName) expression in quo
}
\examples{
none.
}
