% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{remove_emptyGeometry}
\alias{remove_emptyGeometry}
\title{Remove sf data frame with empty geometry}
\usage{
remove_emptyGeometry(sf_elections)
}
\arguments{
\item{sf_elections}{a sf data frame}
}
\description{
When a sf data frame has rows of empty geometry, ggplotly would produce error.
}
\examples{
none.
}
