% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{reorderTraceByLegendgrouptitles}
\alias{reorderTraceByLegendgrouptitles}
\title{Reorder legend groups based on desired group titles in order}
\usage{
reorderTraceByLegendgrouptitles(p, legendgrouptitles)
}
\arguments{
\item{p}{a plotly plot with legend group titles already there.}

\item{legendgrouptitles}{a character vector of legend group titles in desired order}
}
\value{
a plotly plot
}
\description{
Reorder legend groups based on desired group titles in order
}
