% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rgba2hex}
\alias{rgba2hex}
\title{Convert RBGA or RBG to hex}
\usage{
rgba2hex(...)
}
\arguments{
\item{...}{A character of "rgba(168, 207, 246, 0.05941)", or inputs of rgba numbers, e.g. 168, 207, 246, 0.05941}
}
\value{
a hex code
}
\description{
Convert RBGA or RBG to hex
}
\examples{
"rgba(168,207,245,0.6941)" |> rgba2hex()
rgba2hex(168, 207, 245, 0.6)
}
