% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{saveBrowse}
\alias{saveBrowse}
\title{Save and browse plotly}
\usage{
saveBrowse(plt0, file = "plt0.html")
}
\arguments{
\item{plt0}{a plotly object}

\item{file}{html file name, default="plt0.html"}
}
\description{
Save and browse plotly
}
