% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ScaleOff}
\alias{ScaleOff}
\title{Generate scale_x/y that removes margins on the axes.}
\usage{
ScaleOff(...)
}
\arguments{
\item{...}{scale_x, scale_y functions}
}
\value{
a list of scale function using which will remove the axes.
}
\description{
Generate scale_x/y that removes margins on the axes.
}
