% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_col2}
\alias{show_col2}
\title{scales::show_col's revision where labels can be text other than color codes.}
\usage{
show_col2(colours, labels)
}
\arguments{
\item{colours}{characters of color}

\item{labels}{characters of labels}
}
\value{

}
\description{
scales::show_col's revision where labels can be text other than color codes.
}
\examples{
none.
}
