% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{showJson}
\alias{showJson}
\title{Show the JSON as a list from a htmlwidget}
\usage{
showJson(tag)
}
\arguments{
\item{tag}{a htmlwidget that has gone through showTag()}
}
\value{

}
\description{
Show the JSON as a list from a htmlwidget
}
\examples{
tag |> showTag();
tag |> showJson();
}
