% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{showTag}
\alias{showTag}
\title{Creat and view a webpage in http mode}
\usage{
showTag(shinyTag = NULL, js = NULL, css = NULL, ...)
}
\arguments{
\item{shinyTag}{default=NULL}

\item{js}{default=NULL, a file path to .js file, or a string of js scripts.}

\item{css}{default=NULL, a file path to .css file, or a string of css context.}

\item{...}{other to pass to tagList}
}
\value{

}
\description{
Creat and view a webpage in http mode
}
\examples{
none.
}
