% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{showWidget}
\alias{showWidget}
\title{As browseURL but it browse a shiny.tag or html class object in RStudio Viewer}
\usage{
showWidget(tag = .Last.value)
}
\arguments{
\item{tag}{A shiny.tag or html class object}
}
\description{
As browseURL but it browse a shiny.tag or html class object in RStudio Viewer
}
\examples{
none
}
