% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{style_groupTitles_basedOnRankedSplitNames}
\alias{style_groupTitles_basedOnRankedSplitNames}
\title{Add legend group titles based on ranked split names}
\usage{
style_groupTitles_basedOnRankedSplitNames(.plot, from, to)
}
\arguments{
\item{.plot}{a plotly plot}

\item{from}{a starting character number for extracting group titles from get_ranked_splitNames(.plot)}

\item{to}{a ending character number for extracting group titles from get_ranked_splitNames(.plot)}
}
\description{
Add legend group titles based on ranked split names
}
\examples{
none.
}
