% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containerLOT.R
\name{subplot_titles}
\alias{subplot_titles}
\title{Supply titles to the subplots panel produced by plotly::subplot()}
\usage{
subplot_titles(
  widget,
  titles,
  id = NULL,
  lower = "40px",
  fontsize = "20px",
  pull = 0,
  leftpull = 0,
  rightpull = 0
)
}
\arguments{
\item{widget}{a widget}

\item{titles}{list or character vector of titles}

\item{id}{default=NULL}

\item{lower}{default="40px", the position titles should be lowered by this amount}

\item{fontsize}{default="20px"}

\item{pull}{default=0, the left and right pull add padding to both sides}

\item{leftpull}{default=0, the left padding}

\item{rightpull}{default=0, the right padding}
}
\description{
Supply titles to the subplots panel produced by plotly::subplot()
}
\examples{
none.
}
