% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{add_legendgrouptitle}
\alias{add_legendgrouptitle}
\title{Add legend group title to a plotly plot that already has legendgroup defined.}
\usage{
add_legendgrouptitle(p, legendgrouptitles = NULL)
}
\arguments{
\item{p}{a plotly object.}

\item{legendgrouptitles}{default=NULL, or a vector of legendgrouptiles in desired order.}
}
\value{
a plotly object. If legendgrouptitles supplied, the plot will have legend groups presented in sequence as in legendgrouptitles.
}
\description{
Add legend group title to a plotly plot that already has legendgroup defined.
}
