% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{getTable_legendgrouptitle_trace}
\alias{getTable_legendgrouptitle_trace}
\title{Get the table of legend group title and trace}
\usage{
getTable_legendgrouptitle_trace(p)
}
\arguments{
\item{p}{a plotly object}
}
\value{
a data frame
}
\description{
Get the table of legend group title and trace
}
